extract OtpType enum
